First pass on Zappers
Made zapper bots which are all controlled by a manager that has them regularly send out tendrils of electricity which, when they meet a tendril from another bot, will "strike". This should eventually harm the player, but this is not yet implemented. Furthermore, the bots are not yet killable and a solution for them to be able to arena zip is being implemented.  The algorithm for deciding which direction the tendrils extend also needs revision, as they currently extend towards the average location of the zapper bots, but should really be drawn towards closer bots more strongly or even to other tendrils specifically. A potential extension of this design is for tendrils to also be drawn towards the player, dealing large damage on a "strike".